+++ NEWS.md
@@ -1,5 +1,6 @@
 # argonR 0.2.0.9000
 - add hex
+- add a view argument to `argonPageTemplate()`.
 
 # argonR 0.1.0
 - first release